+++ eslint.config.ts
@@ -15,6 +15,7 @@ const configs: ConfigArray = tsEslint.config(
 			"**/__temp__",
 			"**/node_modules/",
 			"**/.docusaurus",
+			"specs",
 			"packages/vscode-extension/server/",
 			"packages/nadle/test/__fixtures__/mixed-ts-js/nadle.config.js",
 			"packages/language-server/test/__fixtures__/config.js",